Fixed ResetShield not resetting negated shields; added ResetShield() extension for resetting specific shield abilities
