+++ InscryptionAPI/Triggers/TakeDamagePatches.cs
@@ -110,7 +110,7 @@ public static class TakeDamagePatches
                     // if (HasShield && damage > 0)
                     //   BreakShield();
 
-                    MethodBase breakShield = AccessTools.Method(typeof(TakeDamagePatches), nameof(TakeDamagePatches.BreakShield),
+                    MethodBase breakShield = AccessTools.Method(typeof(ShieldManager), nameof(ShieldManager.BreakShield),
                         new Type[] { typeof(PlayableCard), typeof(int), typeof(PlayableCard) });
 
                     codes.RemoveRange(shieldStart, shieldEnd - shieldStart);
@@ -133,113 +133,6 @@ public static class TakeDamagePatches
                 break;
             }
         }
-        //codes.LogCodeInscryptions();
         return codes;
-    }
-
-    public static void BreakShield(PlayableCard target, int damage, PlayableCard attacker)
-    {
-        // this void assumes that damage > 0
-
-        var components = target.GetComponents<DamageShieldBehaviour>();
-        foreach (var component in components)
-        {
-            // only reduce numShields for components with positive count
-            if (component.HasShields())
-            {
-                component.numShields--;
-
-                target.Anim.StrongNegationEffect();
-
-                // if we've exhausted this shield component's count, negate it
-                // so it updates the display
-                if (component.NumShields == 0)
-                {
-                    target.AddTemporaryMod(new()
-                    {
-                        negateAbilities = new() { component.Ability },
-                        nonCopyable = true
-                    });
-                }
-
-                // if we removed the last shield
-                if (target.GetTotalShields() == 0)
-                {
-                    target.Status.lostShield = true;
-
-                    if (target.Info.name == "MudTurtle")
-                        target.SwitchToAlternatePortrait();
-                    else if (target.Info.HasBrokenShieldPortrait())
-                        target.SwitchToPortrait(target.Info.BrokenShieldPortrait());
-                }
-                target.UpdateFaceUpOnBoardEffects();
-                break;
-            }
-        }
-        var components2 = target.GetComponents<ActivatedDamageShieldBehaviour>();
-        foreach (var component2 in components2)
-        {
-            // only reduce numShields for components with positive count
-            if (component2.HasShields())
-            {
-                component2.numShields--;
-
-                target.Anim.StrongNegationEffect();
-
-                // if we've exhausted this shield component's count, negate it
-                // so it updates the display
-                if (component2.NumShields == 0)
-                {
-                    target.AddTemporaryMod(new()
-                    {
-                        negateAbilities = new() { component2.Ability },
-                        nonCopyable = true
-                    });
-                }
-
-                // if we removed the last shield
-                if (target.GetTotalShields() == 0)
-                {
-                    target.Status.lostShield = true;
-
-                    if (target.Info.name == "MudTurtle")
-                        target.SwitchToAlternatePortrait();
-                    else if (target.Info.HasBrokenShieldPortrait())
-                        target.SwitchToPortrait(target.Info.BrokenShieldPortrait());
-                }
-                target.UpdateFaceUpOnBoardEffects();
-                break;
-            }
-        }
-    }
-
-    [HarmonyPostfix, HarmonyPatch(typeof(PlayableCard), nameof(PlayableCard.HasShield))]
-    private static void ReplaceHasShieldBool(PlayableCard __instance, ref bool __result)
-    {
-        __result = NewHasShield(__instance);
-    }
-    public static bool NewHasShield(PlayableCard instance)
-    {
-        if (instance.GetTotalShields() > 0)
-            return !instance.Status.lostShield;
-
-        return false;
-    }
-
-    [HarmonyPrefix, HarmonyPatch(typeof(PlayableCard), nameof(PlayableCard.ResetShield))]
-    private static void ResetModShields(PlayableCard __instance)
-    {
-        foreach (var com in __instance.GetComponents<DamageShieldBehaviour>())
-        {
-            com.ResetShields();
-            CardModificationInfo mod = __instance.TemporaryMods.Find(x => x.negateAbilities != null && x.negateAbilities.Contains(com.Ability));
-            __instance.TemporaryMods.Remove(mod);
-        }
-        foreach (var com in __instance.GetComponents<ActivatedDamageShieldBehaviour>())
-        {
-            com.ResetShields();
-            CardModificationInfo mod = __instance.TemporaryMods.Find(x => x.negateAbilities != null && x.negateAbilities.Contains(com.Ability));
-            __instance.TemporaryMods.Remove(mod);
-        }
     }
 }
